refactor : httputils 패키지와 302코드로수정
- HttpUtils가 http패키지로 되어있던 문제 수정

- 303코드를 302코드로 수정
